More tests. Added dev\common to various test projects' Include path to fix compile error. Cleaned up (optimized) several test projects.

+++ dev/Interop/StoragePickers/PickerCommon.cpp
@@ -432,7 +432,7 @@ namespace PickerCommon {
 
             if (FocusLastFilter)
             {
-                check_hresult(dialog->SetFileTypeIndex(FileTypeFilterPara.size()));
+                check_hresult(dialog->SetFileTypeIndex(static_cast<UINT>(FileTypeFilterPara.size())));
             }
         }
     }
